Updated item packets to NA233
Pocket was changed from a byte to an int, and since this update the
client doesn't seem to ignore incorrect variable types anymore and
simply ignores those packets, resulting in errors and players getting
stuck.

+++ src/Mabi/Const/Items.cs
@@ -11,7 +11,7 @@ namespace Aura.Mabi.Const
 	/// <remarks>
 	/// Every item is stored in a specific pocket.
 	/// </remarks>
-	public enum Pocket : byte
+	public enum Pocket : int
 	{
 		None = 0,
 		Cursor = 1,
